+++ test/test.dart
@@ -10,6 +10,7 @@ part 'roman_numerals_test.dart';
 part 'turkish_numbers_test.dart';
 part 'credit_card_verifier_test.dart';
 part 'markdown_markup_test.dart';
+part 'gravatar_image_test.dart';
 
 main() {
   initPolymer().run(() {
@@ -18,6 +19,7 @@ main() {
     turkish_numbers_test();
     credit_card_verifier_test();
     markdown_markup_test();
+    gravatar_image_test();
   });
 }
 
@@ -30,9 +32,9 @@ class NullTreeSanitizer implements NodeTreeSanitizer {
 
 
 class PageComponent {
-  final PolymerElement component;
+  var component;
 
-  const PageComponent(this.component);
+  PageComponent(this.component);
 
   Future flush() {
     Completer completer = new Completer();