+++ CHANGELOG.md
@@ -10,3 +10,9 @@
 * Changed &lt;turkish-number&gt; polymer element number attribute from String to int
 * Changed &lt;creditcard-verifier&gt; polymer element to display red border on invalid credit card number and from blue border, no border if correct
 * Created &lt;markdown-markup&gt; polymer element
+
+## 0.2.0
+
+* Made markdown_markup.dart into a library fun_components.markdown_markup
+* Created a &lt;gravatar-image&gt; polymer element where you can specify the size of the image to use.
+